+++ models/src/main/java/io/sphere/sdk/products/commands/updateactions/ChangePrice.java
@@ -18,7 +18,7 @@ public class ChangePrice extends StageableProductUpdateAction {
         this.price = price;
     }
 
-    public long getVariantId() {
+    public int getVariantId() {
         return variantId;
     }
